refactor(customer): reorganize mappers and improve type safety
- remove separate mapper index files and consolidate exports
- replace persistence interfaces with domain types
- update zod schema to use uuidv7
- make pagination params optional in DTOs
- add documentation for repository methods
