+++ src/domains/customer/aggregates/entities/index.ts
@@ -1,2 +1,8 @@
-export * from './customer.entity';
-export * from './customer.attributes';
+export { Customer } from './customer.entity';
+export {
+  ICustomerBase,
+  ICustomerCreate,
+  ICustomerEntity,
+  IWishListCreated,
+  ICustomerReviewCreated,
+} from './customer.attributes';
