+++ src/domains/customer/aggregates/value-objects/index.ts
@@ -1,1 +1,2 @@
 export { WishListItem } from './wish-list-item.vo';
+export { CustomerReviewProduct } from './customer-review-product.vo';
