--- src/domains/customer/application/mappers/customer/index.ts
@@ -1,2 +1,0 @@
-export * from './customer.dto';
-export * from './customer.mapper';
