+++ src/domains/customer/application/mappers/index.ts
@@ -1,1 +1,8 @@
-export * from './customer';
+export { CustomerDTO, PaginatedCustomersDTO } from './customer/customer.dto';
+export { CustomerMapper } from './customer/customer.mapper';
+export { WishListMapper } from './wish-list/wish-list.mapper';
+export {
+  WishListDTO,
+  WishlistItemWithVariantDTO,
+  PaginatedWishlistDTO,
+} from './wish-list/wish-list.dto';
