--- src/domains/customer/application/mappers/wish-list/index.ts
@@ -1,2 +1,0 @@
-export * from './wish-list.mapper';
-export * from './wish-list.dto';
