+++ src/domains/customer/application/queries/many/review/find-many-customer-reviews.dto.ts
@@ -2,7 +2,7 @@ export class FindManyCustomerReviewsDto {
   constructor(
     public readonly customerId: string,
     public readonly reviewIds?: string[],
-    public readonly page: number = 1,
-    public readonly limit: number = 25,
+    public readonly page?: number,
+    public readonly limit?: number,
   ) {}
 }
