+++ src/domains/customer/application/queries/many/wish-list/find-wish-list-items.dto.ts
@@ -2,7 +2,7 @@ export class FindWishlistItemsDto {
   constructor(
     public readonly customerId: string,
     public readonly variantsIds: string[] = [],
-    public readonly page: number = 1,
-    public readonly limit: number = 25,
+    public readonly page?: number,
+    public readonly limit?: number,
   ) {}
 }
